# api/docs/swagger.yaml
basePath: /
definitions:
  health.AddLifeStyleDataReq:
    properties:
      data_type:
        type: string
      data_value:
        type: string
      user_id:
        type: string
    type: object
...
  health.GetLifeStyleByIdRes:
    properties:
      data_type:
        type: string
      data_value:
        type: string
      id:
        type: string
      recorded_date:
        type: string
      user_id:
        type: string
    type: object
  health.GetLifeStyleDataByIdRes:
    properties:
      life_style:
        items:
          $ref: '#/definitions/health.GetLifeStyleByIdRes'
        type: array
    type: object
...
  health.UpdateLifeStyleDataReq:
    properties:
      data_type:
        type: string
      data_value:
        type: string
      id:
        type: string
    type: object
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.GetLifeStyle:
    properties:
      data_type:
        type: string
      data_value:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      recorded_date:
        type: string
    type: object
  models.GetProfileRes:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
...
  models.Success:
    properties:
      message:
        type: string
    type: object
  models.UpdateProfileReq:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
    type: object
...
  user.FilterUsers:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      role:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: api gateway service
  title: Api Getaway
  version: "1.0"
paths:
...
  /api/lifestyle/addLifestyleData:
    post:
      consumes:
      - application/json
      description: Adds lifestyle data for a user
      parameters:
      - description: Request body for adding lifestyle data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/health.AddLifeStyleDataReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Add lifestyle data
      tags:
      - Lifestyle
  /api/lifestyle/deleteLifestyleData/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific lifestyle data entry by its ID
      parameters:
      - description: Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete lifestyle data
      tags:
      - Lifestyle
  /api/lifestyle/getAllLifestyleData/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieves lifestyle data for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.GetLifeStyle'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get lifestyle data
      tags:
      - Lifestyle
  /api/lifestyle/getLifestyleById/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves lifestyle data for a specific entry by its ID
      parameters:
      - description: Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/health.GetLifeStyleDataByIdRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get lifestyle data by ID
      tags:
      - Lifestyle
  /api/lifestyle/updateLifestyleData:
    put:
      consumes:
      - application/json
      description: Updates a specific lifestyle data entry
      parameters:
      - description: Request body for updating lifestyle data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/health.UpdateLifeStyleDataReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Update lifestyle data
      tags:
      - Lifestyle
...
  /api/user/email/{email}:
    get:
      description: Retrieves a user’s information by their email address
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/user.FilterUsers'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get user by email
      tags:
      - User
  /api/user/profile/{id}:
    get:
      description: Retrieves the profile information of a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.GetProfileRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get user profile
      tags:
      - User
  /api/user/updateUser/{id}:
    put:
      description: Updates the profile information of a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body for updating user profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileReq'
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - User
...
schemes:
- http
swagger: "2.0"